// web/PirateCardsLayaGame/client/Game/src/GameFrame/Menu/MenuLayerType.ts
// ===========================
// 层级类型
// ---------------------------
export enum MenuLayerType
{
    // 主界面
    Home = 1,

    // 主界面 上一层
    HomeTop = 2,

    // 模块
    WarUI = 3,

    // 模块
    Module = 4,

    // 模块
    ModuleTop = 5,

    // 主UI
    MainUI = 6,

    // 对话框
    Dialog = 7,

    // 引导
    Guide = 8,

    // 加载面板
    Loader = 9,

    // GM
    GM = 10,

    // 浮动对话框
    FloatMsg = 11,
}